+++ src/pages/Memo/page.tsx
@@ -11,13 +11,13 @@ import { useAppDispatch } from '@hooks/useRedux';
 import { useIsMobile } from '@hooks/useWindowSize';
 import MemoComponent from '@components/Memo/Memo';
 import MemoBottomButtonMenu from '@components/Menu/MemoBottomButtonMenu';
-import AddImageMemo from '@components/Modal/AddImageMemo/AddImageMemo';
-import AddScheduleMemo from '@components/Modal/AddScheduleMemo/AddScheduleMemo';
 import ModalLayoutGray from '@components/Modal/Layout/ModalLayoutGray';
 import MemoCollection from '@components/Modal/MemoCollection/MemoCollection';
 import MemoDetailImage from '@components/Modal/MemoDetail/Image/MemoDetailImage';
 import MemoDetailSchedule from '@components/Modal/MemoDetail/Schedule/MemoDetailSchedule';
 import MemoDetailText from '@components/Modal/MemoDetail/Text/MemoDetailText';
+import AddImageMemo from '@components/Modal/MemoEditor/Image/AddImageMemo';
+import AddScheduleMemo from '@components/Modal/MemoEditor/Schedule/MemoEditorSchedule';
 import SearchMemo from '@components/Modal/SearchMemo/SearchMemo';
 import CalendarSVG from '@assets/icons/calendar.svg?react';
 import PaperClipSVG from '@assets/icons/paperclip.svg?react';
